drv/core: move enable interrupts to driver init
On hubs with the USB/DFU bootloaders, interrupts are disabled by the
`cpsid` instruction in the bootloader before jumping to the main
firmware. We were reenabling them in early boot. However, this caused
the Essential hub to hang if you press and hold the button to activate
the bootloader without connecting USB then release the button after
the lights start flashing. This is fixed by moving the enable until
after all of the driver init functions have run as in the official LEGO
firmware.

Fixes: https://github.com/pybricks/support/issues/821