drv/bluetooth: Make broadcast handler call consistent.
In the btstack driver, it was called straight from the event/packet handler.
In the cc2640 driver, there was a separate process loop.

The problem with the latter is that it would set advertising_data_received
back to false, which would mess with other scan-and-connect processes, such
as those of the remote.

This will also let us make a finite start_scan task for both platforms.
Before this commit, it would never end in case of the cc2640.

In the future, we might want to fire a generic event to say that new advertising
data is ready for processing. Then the broadcast handler could be called from
a process local to pbio/broadcast.