drv/bluetooth_stm32_cc2640:broadcast random timing
The broadcast process periodically switches between scanning and advertising.
Previously this was done with a fixed timing interval (50ms).
This could lead to this process being fully synchronized, i.e., both hubs are scanning at the same time and advertising at the same time, which makes it imposible to transfer data via the broadcast.

This commit introduces a pseudo random timing of both the scanning and advertising interval. Each in the interval [15,35] ms.
This makes it virtually impossible to have fully synchronization for multiple periods in a row.