pbio/servo_settings: Move friction torque reduction.
The friction torque value was previously reduced because friction compensation caused too much overshoot. But the model isn't the right place to do this.

This restores the original values and moves the factor 2 reduction to wherever the friction value is used. Now, we can decide for each case if it makes sense to keep this reduction.