pybricks/util_mp/pb_obj_helper: fix crash caused by PB_ATTRIBUTE_TABLE
Using the key of the locals dict to store non-qstr info causes the
builtin help() method to crash. Instead we can use the protocol slot
to avoid the crash. The attr entry arrays are now zero terminated
instead of storing the size (since the entries are < 32-bits we are
not using any more space than we would if we stored the size somewhere
else where it had to be 32-bit aligned).

Fixes: https://github.com/pybricks/support/issues/909