pybricks/util_pb/pb_device_stm32: fix pb_device_set_power_supply port offset
This was using the I/O port offset instead of using the motor port
offset. Since it controls the motors, it needs to use the motor port
numbering.

This was only noticed on move hub since it is the only hub where there
are motor ports without I/O ports.

Fixes: https://github.com/pybricks/support/issues/913