pbio/control: Fix logging of speed integral reference.
In speed control mode (run forever), we use an adjusted "position" value to obtain an exact integral of the speed error. This adjustment was not being logged, which made it harder to see what the P action was doing. This fixes it by re-adjusting the value to be logged in application units.