pbio/control: Use reference speed direction to decide stall.
This is more deterministic than using the speed error, which is state dependent. Furthermore, we were using the speed error via the observer, which does not always have the correct direction under load.

This caused the trajectory pause condition to be delayed. In turn, on release, this meant it needs to catch up more which contributes to a larger overshoot on release. This reduces it.