github/build-stats-web: only include mainline commits
The commit tree has several grafted trees that we don't need to include
so ensure each commit in the graph is one from the main ancestor.

Also improve the min/max selection while we are touching this so plots
don't have data points on the top and bottom edges of the graph.