platform/nxt/nxos: move interrupt disable/enable to C code
This moves the NxOS interrupt enable and disable functions to C code.

This is mainly needed for proper attributes on these functions so we
can enable -mthumb -mthumb-interwork at a later time.

While we are touching this, we change the API to use a local state
variable instead of a global enable count which makes it fit better
with similar functions on other platforms.

The assembly file is renamed to irq.S to avoid an .o file name clash
with the new .c file. It now only contains the interrupt handlers.