pbio/drv/sound: new driver for NXT
This implements a new sound driver for NXT. The NxOS sound driver did
not support PCM playback, so we copied the lejos driver instead. It is
modified to work with the pbdrv_sound APIs.

TODO: It currently doesn't work well with the Pybricks MicroPython
speaker class since the speaker class uses extremely high bit rates.
Also the fact that the playback is rounded to a certain number of
samples creates dead spots in repeating sounds that are supposed to
be continuous tones.