pbio/drv/imu_lsm6ds3tr_c_stm32: use INT1 for data ready
This changes the LSM6DS3TR-C driver to use the INT1 signal to indicate
when data is ready to be read instead of reading as fast as possible.
This reduces CPU usage quite a bit since we now have some down time
between reading samples.

Also remove temperature reading while we are touching this. Experiments
show that temperature doesn't make a measurable difference in readings
so we don't need it and it will further reduce CPU usage.

Fixes: https://github.com/pybricks/support/issues/718