micropython: STM32F4 I2C fixes
MicroPython has an old (2017) version of the STM32F4 HAL which has some
bugs in the I2C driver that have been fixed in more recent versions.

This pulls in a change that includes a Pybricks fork of the MicroPython
stm32lib repository that includes the updated I2C driver.

Also include a new tool for tagging releases of our fork of the stm32lib
repository.

Fixes: https://github.com/pybricks/support/issues/932