pbio/control: Use measured speed error to decide stall.
This partially reverts commit fcccd11e9b8ca414e252489a487bd70eaa7a22d0.

When using only the speed reference, there can be a situation where P is fully saturated
when D is opposite and balancing it out. But this means that it is already moving, so the
pause condition should be invalidated.

The commit is not fully reverted, because now we use the measured speed error.
The issue of pausing is only relevant under load, in which case the measured speed
falls more quickly than the estimated speed (which doesn't experience the load).

Fixes https://github.com/pybricks/support/issues/946