pbio/servo: Handle CONTINUE as HOLD in stop.
This method is not exposed to the user so they cannot
accidentally provide the CONTINUE argument to this function.

But it could be called implicitly when reset_angle is used.
Now the stop function at least stops the motor for all inputs.

So we cast CONTINUE to HOLD. We use HOLD instead of coast
so that this is not a breaking change. Previously, doing reset
during active movement also resulted in hold.