pbio/integrator: Simplify position integrator state.
This changes the used integral value to be one-step-ahead instead of
returning the previous error. This makes little difference in the overal
output (it marginally changes the proportional term), but makes it
easier to adjust the range limits on the current error instead of the
previous error. That is, the position error and the target error are
now from the same time stamp.

Also removes one state value which was no longer used.