pbio/trajectory: Store user given speed.
This value can be used during control to test if the original speed target
was low. This can be used to reduce the control gains to increase
stability as a tradeoff against stiffness.

We use the user given speed instead of the actual top speed, so that
this won't affect commands that have low just only because the
maneuver is short.