pbio/control: Compensate reference for stalling.
When transitioning from a time-based trajectory to a position-based
maneuver, the built up integration error was not accounted for. This
caused an instantaneous jump in the reference position when, for
example, running for time under load and then running to a position.

The reverse case (from position to timed) was already accounted for
long ago by compensating for time spent stalling.

Fixes https://github.com/pybricks/support/issues/956.