drv/bluetooth_stm32_cc2640: propagate all events
Not all of the event handlers were triggering the task queue which
could cause a delay while waiting for a response. If a task was waiting
for an event from a handler that did not run tasks, then it would have
to wait for another event before continuing.

This caused a very long delay when connecting to the Powered Up remote
when the hub was not connected to a computer at the same time (the
hub sends status notifications ever 500ms so the delay was not as
noticeable in the connected case).

Fixes: https://github.com/pybricks/support/issues/466