bricks/virtualhub: Build without CPython.
This uses the test drivers instead of the CPython implementation.

This is easier to debug and maintain, and it runs a lot faster.

No specialized configuration is needed to run it anymore. You can just run the virtualhub-micropython executable as if it is normal MicroPython. It just so happens to have a working `pybricks` package.

The robot setup is defined in platform.c. It currently uses 5 motors, but this can be changed later to simulate a few sensors.

Lights and buttons are not done yet, which is a regression for now.