pbio/servo_settings: Round rated speeds.
The rated speeds for the technic motors were obtained from the
information coming from the motor. They are indicators of
what the motor can run at under nominal circumstances without load.

Since this is a very arbitrary measure, we might as well round them by
a few percent to give them values that are easy to remember. This
makes both angular motors work the same, and both linear motors
work the same.