pbio/servo: Refactor settings loader.
This cleans up and moves the device-indepent settings to pbio/servo.

Also refactor to set both settings structures in one struct literal to
make it easier to verify that all settings are set, indepently.

This allows us to add a user call to reload all settings with a given
precision profile.