pbio/src/integrator: fix parameter direction
`[in,out]` would mean than the when the function returns, `*itg` could
point to a different instance from the one that was passed in (only
possible if passed as `**itg`). So this is an `[in]` parameter. The fact
that `*itg` does not have a `const` qualifier tells us that the instance
state could be modified.