pybricks.robotics.DriveBase: fix pb_attribute_handler guard
When the control attributes are disabled (e.g. on move hub), we must
also disable the attribute callback, otherwise trying to read an
attribute will attempt to read from an array starting at memory address
0. On move hub, this is actually a valid memory address (can be mapped
to flash or ram), so it could cause a crash or not depending on what is
at that address.