drv/bluetooth_stm32_bluenrg: strict ordering of BLE tasks
This reworks the BlueNRG driver to strictly order tasks instead of
running them in parallel. This ensures that each task is complete
before starting the next to try to help avoid confusing the Bluetooth
chip by telling it to do too many things at the same time.

One possible drawback of this is that the scan and connect function
can be very long running so this means the 500ms status updates will
be blocked during this call. However, we can consider alter solutions
if this turns out to cause problems.