pybricks.common.IMU: Change threshold setter.
This makes the method consistent with other settings getters/setters in Pybricks.

In pbio, these values are moved to the config struct. After all, they are config values meant to be changed at runtime, and this lets us avoid having a bunch of setters/getters at multiple layers that do nothing but pass values along.