drv/bluetooth: add broadcast/observer APIs
This adds new Bluetooth driver APIs to enable broadcasting and
observing.

Technically speaking, we aren't using the BLE Broadcaster Mode or
Observer Mode since those are not available on all of the Bluetooth
chips. However, the end result is essentially the same - the same
data is going over the air.

There appears to be a firmware bug in the Bluetooth chip on the city
hub that prevents broadcasting non-connectable advertisements from
being transmitted over the air even though the commands succeed without
error, so this is noted in the code comments.