pybricks.robotics: Add GyroDriveBase.
This is a variant that uses the gyro for steering.

This makes it easier to configure compared to specifying even more
kwargs. This also lets you toggle between drive bahaviors by just
changing the class name in your script, so you can determine what
works best for a robot.

Since this can be completely disabled on hubs without a gyro, the reduced number of parameters also reduces move hub build size a bit.