pbio/drivebase: Use gyro for speed state.
While we have to use the motor speed estimate to get stable motor
motion, we can still use the measured gyro rate for things like stall
detection.

This improves accuracy of stall detection when the robot is lifted
in the air. This is not a typical operation that happens in the field,
but people often do this to test that the gyro is "working",
so we should make sure to produce intuitive behavior.

Partially fixes https://github.com/pybricks/support/issues/1032