pybricks/tools/pb_type_awaitable: Refactor.
This restores the ability to correctly cancel parallel awaitables
that belong to the same resource. This was removed while unifying the
awaitable types in previous commits.

In practice, this means that if a motor operation is awaited while
another motor operation is already running, the older operation will be
cancelled and the new one will start.

Flags are available to control the cancel behavior so the same code can
be shared with awaitables that don't need linking like wait.