pybricks/tools/pb_type_awaitable: Generalize return value.
In many cases, the await part is the same across methods and only the return value differs. This generalizes the awaitable by splitting the completion test and return value.

This will be particularly useful for sensors later on. All methods await mode switches in the same way, but they have different return values.