pybricks/tools/pb_type_awaitable: Generalize object type.
This allows us to await processes without going through an
explicit MicroPython object. This doesn't make a difference
for the awaitables implemented so far, but this will be useful
for Bluetooth and sensor awaitables.

For example, for pupdevices the void *object can be the *iodev
so that the same awaitable type can be used for all sensors,
even though each of the sensor_obj_t are all slightly different.