drv/bluetooth: handle remote + observing
If observing has already started when connecting to another Bluetooth
device, we need to pause observing (which is passive scanning) so that
we can do active scanning to find the device we want to connect to.
Then we need to restore passive scanning for observing once active
scanning is complete.

This is done on CC2540 and BTStack. On BlueNRG, the chip can't handle
observing while being connected to another device so we just raise
an error instead.