pbio/iodev: Refactor uartdev interface.
The current iodev and uartdev code will only ever be used for LEGO uart
devices, so we can reduce the level of abstraction and complexity
considerably.

Combined with the previous and next commits, this reduces code size by
about 580 bytes and helps prepare for async functionality later on,
which is the main reason for this refactoring.

The goal it to keep src/uartdev.c completely in charge of uart devices,
instead of relying on user modules correctly calling operations like
set_mode_end.