drv/bluetooth_stm32_cc2640: add city hub NONCONN_IND quirk
The Bluetooth chip firmware on city hub fails to send anything over the
air when using the NONCONN_IND advertisement type. This adds a hack to
use ADV_IND instead which is a connectable advertisement. Things will
break if someone actually connects but this is better than not being
able to broadcast.

It also means we can no longer filter for NONCONN_IND in the observer
which means we will pick up advertisements from the official LEGO
Robot Inventor firmware. Most of these will be ignored since the
first byte of the CRC32 in the LEGO data is unlikely to match any of
the observed channels.