pbio/drv/legodev_pup: Add boot power quirk.
The city hub powers on the I/O ports (pin 4) during boot, so by the
time the device connection manager detected a UART device, in many
cases, it was too late to catch the device on the first sync. In the
case of the BOOST color distance sensor, this meant waiting 10 seconds
instead of 5 for the sensor to be ready.

Something similar was happening on Technic Hub, where it would always
miss the first synchronization cycle.

This adds a workaround that cycles power to the I/O ports immediately after
boot. This is faster than waiting for the next sync cycle.

Issue: https://github.com/pybricks/support/issues/747