drv/bluetooth_stm32_cc2640: work around broadcast lock up
The Bluetooth chips on the Technic and City hubs seem to lock up if
the advertisement date is updated too many times. So far, we have only
found one condition where this doesn't happen. The Bluetooth chip
can't be connected to any other devices and advertising has to be
stopped and restarted between each update. This is quite limiting,
but seems to be the best we can do for now.

Issue: https://github.com/pybricks/support/issues/1095