drv/bluetooth_stm32_cc2640: fix broadcast lockup
This changes the command to update advertising data on the Bluetooth
chip in the broadcast task to use HCI_LE_setAdvertisingData() instead of
GAP_updateAdvertisingData().

It appears there is some sort of memory leak or some such on the
Bluetooth chip that causes it to lock up if we call the vendor-specific
GAP_updateAdvertisingData() too many times. It appears that we can work
around this by calling the standard Bluetooth command
HCI_LE_setAdvertisingData() instead.

GAP_updateAdvertisingData() would result in two events being sent back
from the Bluetooth chip, a status event and a command complete event.
With HCI_LE_setAdvertisingData() we only get a command complete event
back - with the TI vendor-specific opcode as if we had called
GAP_updateAdvertisingData().

Fixes: https://github.com/pybricks/support/issues/1095