drv/bluetooth_stm32_cc2640: enable connect whitelist when broadcasting
This changes the GAP discoverable parameters to limit connecting to
devices in a whitelist. Normally, when broadcasting, the hub should
not be connectable, but due to some quirks of the Bluetooth chip,
ADV_NONCONN_IND doesn't work under all circumstances and we have to use
ADV_IND which makes the device connectable. If another device connects
while broadcasting, it currently causes things to lock up. However,
we can prevent devices from connecting in the first place by enabling
the whitelist (which is empty).