drv/bluetooth_btstack: finish TODO to wait for setting advertising data
This finishes up the REVISIT comment in start_broadcasting_task() to
actually wait for the set advertising data command to complete before
the task completes.

This fixes issues with calling the task too fast while the commands
from the previous call are still pending.

The timer was used to work around the issue but is no longer needed.

Fixes: https://github.com/pybricks/support/issues/1151