pybricks.common.ble: fix data consistency when observing
In the `ble` module, we maintain a copy of observed advertisement data
that is updated during servicing of PBIO (contiki) events. These events
can be triggered any time the GC is invoked, e.g. during allocation of
memory for MicroPython objects.

In pb_module_ble_observe(), we were iterating over the observed data
and allocating new MicroPython objects for each item which could trigger
PBIO events to be run resulting in the observed data being modified
while we are still iterating over it. This could result in a crash, the
most likely of which is raising an exception for an invalid data type.

We can avoid this by making a copy of the data before iterating over it.

Fixes: https://github.com/pybricks/support/issues/1278