lib/STM32_USB_Device_Library: Replace with submodule
Replaces the copy of the STM32 USB library with a submodule.
Uses v2.6.0 as that was roughly the version that was in the
repo. Temporarily adds #undef for MIN and MAX to get it to
compile. Removes the CDC file from the build as a subsequent
change will replace that with a custom class.

Co-authored-by: David Lechner <david@pybricks.com>
Signed-off-by: Nate Karstens <nate.karstens@gmail.com>