drv/bluetooth_stm32_cc2640: fix ble observe
The Bluetooth chip on the City and Technic hubs will stop receiving
advertisements after a while when observing. It isn't clear why this
is happening, but it seems to that there is a bug in the Bluetooth
chip firmware that causes it to start filtering out advertisements
from an individual advertiser after a while (i.e. it can still be
receiving advertisements from one device but stops receiving them
from another). Eventually it will stop receiving advertisements from
all devices.

To work around this, we restart the observation process every 10
seconds. In testing, we were never able to trigger the bug in less than
25 seconds or so. Therefore, 10 seconds should be more than enough to
ensure that we never miss an advertisement.

Fixes: https://github.com/pybricks/support/issues/1096