drv/bluetooth: add task parameter to pbdrv_bluetooth_stop_observing()
This adds a task parameter to pbdrv_bluetooth_stop_observing(). This
allows us to partially fix a TODO of waiting for the observation to
stop before exiting the MicroPython runtime. Also in the case of the
CC2640, pbdrv_bluetooth_stop_observing() can be called from multiple
places, so to be safe, we need to make sure each caller has its own
task struct to avoid reentrancy issues.