drv/usb/stm32_usbd: use Pybricks class/subclass/protocol
This adds new defines for the Pybricks USB device class, subclass and
protocol. The class (0xff) is defined by the USB-IF and is the vendor-
specific class. The subclass and protocol are vendor-specific and come
from the first two bytes of the the Pybricks Bluetooth UUIDs.

This allows us to identify hubs running Pybricks when enumerating USB
devices.