pbio/drv/bluetooth: Generalize peripheral state.
This takes it to the driver, where the start of the scan task sets the relevant callbacks.

Also generalize names like pbdrv_bluetooth_write_remote to pbdrv_bluetooth_peripheral_write.

Also rename "context" to "peri" (peripheral) where relebant so that things like peri->bdaddress are immediately clear which bdaddress it is.