drv/bluetooth/stm32_cc2640: send correct max size
This finishes a TODO in the CC2640 driver to send the correct max
payload size to the host via the Pybricks hub capabilities
characteristic. Previously, this was hardcoded to use the hub max MTU
size. But some Bluetooth adapters have a smaller MTU size, so we need
to use the smaller of the two.

Fixes: https://github.com/pybricks/support/issues/947