drv/bluetooth_stm32_cc2640: respond to prepared write
If a host tries to write more than MTU - 3 bytes to an attribute, the
host should do a prepared write to split up the write into multiple long
writes. We currently don't support this. So we need to send an error
response to the prepared write request. The host may also follow this
with an execute write command to cancel the prepared write, so we need
to respond to that too.

Issue: https://github.com/pybricks/support/issues/947