pybricks.common: Dynamically create button enum.
This simplifies the button enum to make them more suitable beyond just pbio-specific keypad mappings. This will let them be used for other devices, like the Xbox Controller.

This also saves about 200 bytes on Move Hub. Some of this comes from omitting buttons that can't be used on Move Hub or with the remote anyway.