pybricks.robotics.Car: Raise error on no car built.
We get quite a few questions about the Xbox controller not connecting,
when the actual issue is the car initialization never completing because
no car has been built yet. To be more explicit, raise an error if the endpoint
is not found within 10 seconds.

Fixes https://github.com/pybricks/support/issues/1564