pbio/drv/bluetooth: Finish implementing remote disconnect.
This was started as a task on some platforms, and waiting for completion was handled differently for each platform.

This uses a task so we can cleanly finalize this task along with broadcasting clean up in the following commits.

Now we could also implement an awaitable disconnect method on the Remote.