pbio/protocol: Don't use status flag for BLE enabled.
This reverts 8c8340b6814c15a74a5ed1040cfe7ebc241f57fd.

At this point, we don't need an extra status flag. Bluetooth is the only communication method, and we can't communicate a status if it is disabled.

So we can avoid changing the protocol for now. We can add appropriate flags if we add USB support, in which case communicating the active connection type may make sense.