pbsys/main: Generalize and clean up program start commands.
This brings the waiting process for the main program to start to the
main.c loop, which makes it easier to follow.

This also generalized the commands to start builtin and user programs
with an optional identifier. For now, this will be used to start either the
REPL or Port View. In the future, it can be used to start programs on
different slots.

This also adds a hook for verifying the program at the application level
so we can verify MicroPython-specific validity and avoid hardcoding such
checks at the pbio level.

pbio/sys/config: Consolidate app config.

The separate placement of these settings was a bit contrived. We can have it
at the pbsys config level since the definitions exist at the pbio protocol level anyway.

This lets us use them as defined feature guards.