pbsys/bluetooth: Do not use bluetooth light indicator.
The main status light not blinking is enough of an indicator for now that Bluetooth is off. This way we don't have to introduce new UIs yet that get locked in the upcoming release.

When we do create a new UI, we can now use the status flags instead of poking at the lights directly from the sys/bluetooth driver.