pbio/sys/light: Restart animation on program start.
This ensures that the animation looks consistent
when you start a program, rather than picking up
where it stopped last time.

This makes it easier to see that a program was
started when you run it while connected, since the
light always switches from solid to off and then
fades.