pybricks/tools/pb_module_tools: remove gc_collect() from run loop
gc_collect() can take a long time to run (milliseconds) which means
that it isn't possible to get smaller loop times than that.

This was here in the first place to help prevent fragmentation, but
there are alternative solutions, like setting gc.threshold() or manually
running gc.collect() in an application that can handle the delay.