pbio/sys: move VCC power off to after shutdown
Up to now, the VCC power off was done when shutdown was requested. This
will cause external sensors to power off. Users may see this and assume
that the hub is now off. However, the hub is still busy with the
shutdown process, e.g. writing to flash memory. So we don't want the
user to release the power button yet.

Instead, move the VCC power off to after the hub has finished everything
and it is safe to power off.

This lets us remove the sys/io_ports.c and sys/io_ports.h files. The
function is renamed to pbdrv_ioport_power_off() to be consistent with
pbdrv_reset_power_off().