pbio/drv/ioport/ioport_pup: add PBDRV_CONFIG_IOPORT_PUP_QUIRK_SHUTDOWN
Add a new PBDRV_CONFIG_IOPORT_PUP_QUIRK_SHUTDOWN configuration flag for
the ioport_pup driver. This flag is used to indicate that the driver
that special handling of the I/O port VCC pin is needed during shutdown.

We already had such a quirk in place, but it was enabled at all times
on all hubs. This improves the situation by only enabling the quirk on
hubs that need it. Also, on hubs that needed it, we now only enable the
quirk when the button is not pressed. This has the desired effect of
turning off sensors before the user released the button so that it looks
like the hub is really off. This way it will be more clear to the user
that they can release the button because the hub is "off".