pbio/sys/main: turn off all motor outputs on shutdown
In most cases, pbio_stop_all() was already stopping motors, but it
doesn't turn off motor outputs for non-motor devices that need extra
power. This was causing the SPIKE Light Matrix to turn on all green LEDs
when any hub was turned off.

This fix is a bit of a hack since it is going behind the back of the
higher-level I/O port stuff, but it works for now.