pbio/imu: Fix internal rounding error from imu to motor control units.
The drivebase controller operates in terms of millidegrees of the wheels. When the gyro is used for heading, it is scaled to these wheel units. Depending on the robot dimensions, this scaling could be off by one or two degrees per rotation due to a rounding error.

Fixes https://github.com/pybricks/support/issues/1886