pbio/drv/charger: Move charge enable logic.
The charger module should be in charge of charging, not the battery module.

This ensures that the battery status getter is synchronized with the actual battery state. Before this commit, the battery could be charging before it was reflected in the charging state.

This will also give us more control to disable charging under certain charger error state conditions in the next commit.