pbio/drv/core: Drop process autostart.
This was a longstanding chore that we have done gradually over time.

The ADC driver was the last to be tranformed to the manual start format.

To ensure that this isn't a potentially breaking change in itself, we move the call to initialize the ADC driver to the end of pbdrv_init, which is where the process was previously auto-started.