pbio/drv/uart_ev3: Enable UART via PRU.
This is a rework of the lms2012, ev3dev, and EV3RT drivers for the PRU software UARTS for sensor ports 3 and 4 on an embedded platform without Linux.

When building and changing PBDRV_UART_DEV_ID in mphalport.c to 3, this produces a working MicroPython REPL on sensor port 4.

Further improvements could be made, such as getting rid of one redundant ringbuffer.