pbio/port: Refactor counter logic.
This brings back the counter driver for hubs with on-board quadrature counters.

This also introduces separate indices for external ports and counters so we don't have to assume everything coincides with having UART or not.

The positioning logic for the LUMP motors is pushed to the LUMP code, rather than having it make callbacks to the port. This is simpler, and it does kind of belong there since its angle encoding is LUMP specific.