pbio/drv/counter: Implement EV3 motor detection.
In all three counter drivers, the device detection is closely coupled to the counter, so we include identification in these drivers.

This simplifies runtime disconnection errors considerably as well, consistent with LUMP devices disconnecting.

We can reconsider this model to a more generic quadrature interface or a dedicated motor detection process if we ever need one.