pbio/drv/gpio_ev3: Drop pbdrv_gpio_alt_gpio.
On EV3 it is necessary to set GPIO alt mode before using the pin. It it simpler for external code to just always do this within the gpio get and set functions.

If we need to reduce overhead in high performance applications like GPIO-based I2C, we can instead have dedicated calls to set/get the pin with fewer function calls.