pbio/port_lump: Fix EV3 Color Sensor.
In the previous commit, we removed a grace timeout that allowed LUMP sensors to send 6 bad messages or keep alive timeouts before considering it disconnected.

In most cases, this was giving a long delay when it is better to know that it was disconnected. Apparently, this hack was originally there to suit the needs of the EV3 color sensor, which sends a bad message directly after syncing up, and then would time out if the sensor value didn't change quickly enough to send a new update before the keepalive kicked in.

Another way to take care of that is to give the sensor a kick in some other way, which in this case is to set a different default mode. It still sends the bad message after syncing, but doesn't time out since the new mode will start by sending sensor data.