pbio/port: Drop device kind and type from port state.
It seemed nice at first to have this available but it is better to get the relevant info from the respective module when requested, rather than relying on a potentially incorrect cached value.

Asserting rather than getting type also works better when the type is ambiguous. For example, some NXT analog devices should pass the assertion even if a definitive ID is not available.