pbio/os: Add context to process.
When using the same thread for multiple processes it can be useful to have a reference to the state that belongs to that instance.

This avoids having to use several cases of CONTAINER_OF to get to the parent of a state, which is not always available.

This also brings it one step closer to pbio tasks, which we might model in this way too.