pbio/sys/status: Handle status change synchronously.
This was using broadcast, but there was ultimately only one receiver for this data. We can simplify this and avoid the risk of overrunning the event queue by calling it synchronously, which is safe for this event.

After the refactoring in the previous commit, this is now a trivial change.