pbio/os: Use namespace for macros.
This makes the invocations a bit longer to fit, but it is probably clearer in the long run.

While we are at it, cut one level of abstraction and specify macros in terms of the switch statement directly.

Also, instead of LC (contiki), use the terminology "checkpoint", which is untuitive and also technically quite close to reality