pbio/drv/usb: Disable work in progress in release builds.
USB is not finished yet, so we'll disable the work-in-progress developments until it is ready for beta testing. This lets us proceed with beta releases to test other functionality.

Leaving it enabled without it actually working can cause unexpected lockups such as in lsusb or when writing stdout. Windows users may get notifications about Pybricks USB devices without it actually working or being documented yet.