pbio/drv/bluetooth_stm32_cc2640: Only restart observe when needed.
Everytime we restart observing, there is a small chance of locking up the Bluetooth chip. The restart was added as a workaround for the scan queue getting full so new scan events were not coming in. This usually doesn't happen for a long time, so we can avoid restarting observing until then.

Since there isn't a good way to know if the queue is full, we can request a restart if we are not receiving data when the user is expecting some data. In the worst case, this will still restart it every 10 seconds as it used to do. If all of the subscribed channels get incoming data, restarting is delayed, thus reducing the likelihood of a freeze.