pbio/port_lump: Revert setting EV3 Color Sensor ambient mode.
This was added in 91acfb321b54ac083e00712f087bb2d6f4c51543 to get the sensor going. Since 6c5c1f033504dfbddb91454d4b11bcf0ac7b1bde, we are doing it by sending a keep-alive message, so we don't need to set any mode on this sensor. It seems better to stick to the default reflection mode.