!deleteme: Disable ADC driver.
We still need to add some logic to selectively enable the block device driver and the ADC driver, which both use SPI0 on EV3.

We don't have that yet, so don't let that get in the way of interesting SPI contributions, and just disable the ADC on this development branch.