bricks/ev3: fix entry point address in uImage
The entry point address in the uImage was hardcoded to 0xC0008000, but
the actual entry point address is determined by the linker. It was just
by luck that the entry point was close to the load address and there
weren't any jumps between, so it eventually got to the entry point after
executing some random code.

Instead, we can scrape the entry point address from the ELF file using
`readelf`, which will adapt to any changes. For good measure, we also
scrape the load address, which is also determined by the linker script.